real location of retroarch-joyconfig
i changed the directory to the retroarch-joyconfig binary, which is under the tools directory now. current builds will are pointing to the old location still. if this binary is supposed to be moved/copied over to the old one of /installdir/bin/ then that needs to be changed instead as it is not working that way currently