2Axes 11Keys Game  Pad
This is the "Tomee" gamepad that I have often talked about in the forums. It is cheap and common all on in retail stores and online like amazon. Here is an example of it: https://www.videogamesplus.ca/tomee-snes-usb-controller-snes-stock-usually-ships-within-24hrs-p-20831.html

It is styled like an SNES controller, connects with USB, supports 8-way d-pay