Add experimental mupen64plus (testing) mainline module.
Ricrpi has done a really nice job. But his fork is not updated as frequently as mainline repo. PRs and upstream fixes are not always merged. Mainline repo is working now with rpi. There is slight performance hit, but it is still much faster as mupen64plus-libretro. Moreover the mainline repo is not stripped down for ARMv6 and could run better with PI2. 

Rice is the default video plugin. I added a texture filtering fix to get rid of pixelated textures (shadows, water and so on) and enabled default 2x multisampling.