rework the logic for creating the per system retroarch.cfg files.
If the file doesn't exist, create a new one
If it does, but misses the important #include line, then add it, but do further changes in a .rp-dist file

This avoids overwriting users custom options such as shaders when updating retropie. Users can use the .rp-dist
file to reference any new defaults.