retroarch uses the first setting it finds so when using an #include to include the main config, settings that should override those in retroarch.cfg should be above it.
adjust the per system retroarch.cfg generation, so the #include is placed at the end.

as before if a retroarch.cfg exists, no settings will be changed in it, and instead a retroarch.cfg.rp-dist will be used - however, if the existing retroarch.cfg is missing the all important #include line, it will be added, along with a comment at the top.