Add Colecovision (CoolCV)
I made the romfolder name coleco instead of colecovision because it allows
us to incorporate colecovision for MESS in the future if needs be. Themes and platform are still colecovision for scraping
and themes. easier Joypad configuration is likely to come with the next release of CoolCV for the
Pi with a potential configurable keyboard. Currently exit is right shift + Enter