Create joystickselect.sh
Show the available joysticks and let the user choose what controller to use for RetroArch Player 1-4.

Short description of what this script does:
- puts at the beginning of $configdir/all/retroarch.cfg an "#include" pointing to the file $configdir/all/input-selection.cfg
- let the user manage the input-selection.cfg through dialogs (this file contains the configs for input_playerN_joypad_index for players 1-4).

OBS. 1: the joystick selection doesn't work if the config_save_on_exit is set to true, because the retroarch.cfg is overwritten frequently.
OBS. 2: the depends_joystickselect will put the jslist program at /opt/retropie/supplementary/jslist. This program just list the currently available joysticks in the system and it's a dependency for this script.