Update lr-glupen64.sh
I've done away with the submodules, since it forces pulling in the history of each submodule, which is going to be a lot. So you don't need this anymore when pulling the repo.

You guys can use --depth=1 to speed it up as well but I don't know what the syntax for that would be with the macros you are using.