Create scraper_setup.sh
This scraper_setup bash script will help populate the roms sub-directories with a simple XML file in order to allow the Scraper that is built into the EmulationStation to save files. Even if you have previously created the necessary files for some of your roms, this should create an XML file in each directory that doesn't already have a gamelist.xml.