Add Amiberry Emulator to experimental packages
Based off of UAE4ARM, amiberry is being actively developed. I used the
new installer for UAE4ARM to install this, it has the same shell script
for launching and same default .uae configs for machine type. I’ve
added .uae as a valid rom file type for people with custom settings in
whdload and things like that.

While eI was in platforms.cfg I added the atari800 filetype .car to
Atari 5200 as well, it’s a valid cart type there too.