Request to Add ComicBook_SE Themes
I would like to request the addition of ComicBook_SE Themes for both the 16:9 and 4:3 versions of the theme.  This allows users with Marquee/Wheelart to display them above the video instead of the standard logos.  Due to the current ES restrictions, it is not possible to allow this to happen and still function for users without wheelart/marqueed scraped.