Launch Art Patch
Adds code for a 5th arg to specify launch art. Due to how the script verifies the files are there, I have it trim any extension from the image file path. May act up in the event that you have a path with periods in other places. Originally created this patch as a way to get custom launch art when triggering from, and specified by, an external script.