Full game description in .svm for scraper to hash
This change makes the script write the full game descrption (e.g. "The Secret of Monkey Island (CD/DOS/English)") into the .svm files that ES uses to generate the gamelist.

The purpose of this change is to make the .svm files unique and not empty so that scrapers can hash the .svm files and match them to the actual games. This is how, for example, screenscraper.fr and Recalbox handle ScummVM scraping. 

Attempting to hash on the real game files is problematic because of the number of different variations of files that make up a ScummVM title - files can be modified or compressed and still be valid, resulting in a virtually unlimited number of possible hash values. This change allows each game title to have exactly one hash value.

Someone with some actual bash knowledge should review this; I am a bash newbie.