Updated esthemes.sh with a new vertical only ES theme for mame roms
This theme is designed to be used with your screen rotated to landscape. You will need to add a line of code to the boot/config.txt file. 'display_rotate=3' (or =1 depending on which way you want to rotate). The theme only covers mame/arcade and relevant collections like capcom, data east etc. No consoles, PC or handhelds are covered.