runcommand - add support for rom directories with extension in name
If the given rom is a directory and has an extension in the name, collect
files inside the rom directory with that same extension, then if there is
exactly one file inside the rom directory with the same extension as the
rom directory, use it for the %ROM% token to be replaced.

This behaviour is useful for example in the lr-scummvm emulator to support
launching games directly when selecting the game directory in
EmulationStation if it has a supported system extension.