steamlink: switch to xinit+matchbox session & misc updates
When steamlink is launched without a running Xorg server,
DualShock controllers erroneously act as both a gamepad and virtual
mouse, which causes major input conflicts in-game.

Launching from an Xorg session fixes this issue, but steamlink also
requires a window manager running in order to set the window focus
property (to allow joystick events to be processed).

Additional changes:
* switch to repository version of steamlink package
* update Help text to reflect new onscreen welcome message
* remove overscan override (issue fixed upstream).