added SuperFlappyBirds scriptmodule to libretrocores
Super Flappy Birds is an original-ish game created by me, specifically with retropie in mind. The game and screenshots are [here](https://github.com/IgniparousTempest/libretro-superflappybirds).

It is written as a libretro core, hence why the script is under the `libretrocores` folder, but it installs to the ports folder on retropie. This pattern seems odd, but it is what Dinothawr does, see [here](https://github.com/RetroPie/RetroPie-Setup/blob/master/scriptmodules/libretrocores/lr-dinothawr.sh).