helpers - fix printMsgs() return value
Without an explicit `return` command, functions in bash default to
returning the exit value of the last command executed.

The current implementation of `printMsgs()` uses `[[..]]` tests for
deciding on the output type (dialog, heading, console), however if
the last test (for heading) is false, then this will be the final
exit value to be returned from `printMsgs()`. In other words, for
all cases except "heading", it always will return `1` (false).