image: don't set custom gpu_mem allocation for mesa targets
The RPI VC4/VC6 drivers use the kernel CMA heap allocation for GPU memory
(which defaults to 256MB on 1GB+ targets) instead of the gpu_mem allocation.

Although the gpu_mem allocation is still used, it's needed only for camera and
hardware decoders, but the default allocation (~64-76MB) is sufficient.

Without this change, we are losing ~192MB of user-addressible kernel memory
for no reason.