packages - added the ability to set a package section based on platform flags
The module/package variable rp_module_section used to need a single string containing
the section for the module - eg "core" "main" "opt".

With this change, you can provide additional space separated parameters to specify the
section for a specific platform flag. This is useful for example to have a module by default
in "opt", but then promote it to "main" for certain systems.

eg for mame4all (mostly useful for slower armv6 devices) you could do

rp_module_section="opt armv6=main"

which would place it in the main section only for armv6