Added squiglies around CPU Temperature variables
I added squigly brackets around the CPU Temp variables in the output string.  This is what made the display work correctly for me on my latest version of Raspian on a Raspberry Pi 4.  I suspect that BASH was trying to use the temperature symbol as part of the variable name.