helpers - we need to check for __has_binaries and choose binary/source install for custom packages
this was done on the old getDepends and we need to keep this until the sdl1/sdl2 modules are improved

the sdl1/sdl2 modules have an install_bin which is called even if no binaries for a platform are available.
until we have some mechanism for a module to have it's own hasBinary type function we need to keep a check here.