redream - symlink new states folder
The latest version of redream copies "saves" to "states" then creates a new saves folder. After doing so it
sets a "savesfixed=1" in it's config. With the creation of our symlink for states the copy won't take place,
but I felt it was safest to avoid trying to do this automatically. Users can copy the configs to the new location
manually. It's an experimental module so this seems reasonable and it avoids additional logic and possible issues.