mupen64plus - rework logic including rpi aarch64 support and merging vero4k/mali
There has been a lot of contributions to this module, and unfortunately some of the logic has become rather obfuscated,
with a lot of duplication. This is a combination of an earlier changeset I had, but adding aarch64 needed enough logic
changes, it made sense to include it all. There are more improvements needed, but this de-duplicates some of the logic.
Vero4k may need a revisit.

Changes include:
 * temporarily disable rice and gles2n64 for aarch64 due to build errors (just HOST checking errors so may well be easily fixed)
 * remove a lot of vero4k specific blocks in favour of the mali flag
 * enable mali for module (builds ok on odroid xu)