gpsp / lr-gpsp - use gcc-6 to workaround segfault with gcc-8 on buster
lr-gpsp:

cpu_threaded.c built with -O2 on gcc 8 produces a non working binary. This source already had
a different optimisation level from the rest of the code (O2 vs O3). However, on gcc-8 it causes
a non working binary when built with anything other than O0, but this could have some performance
issues (would need testing on a rpi1).

I compared all the different optimisation flags enabled by O1 and also with any changes from gcc-6,
but not all GCC optimisations are controllable from flags. It was not possible to produce a working
binary with O1 - even disabling all the configurable optimisations enabled by O1.

The issue is most likely undefined behaviour in this file (which has a lot of large macros), and it
probably needs some fixes. However it could be a bug in the gcc-8 optimiser code.

As the issue isn't present with gcc-6, and as gcc-6 is available on Debian buster, the easiest
workaround is to just force the code to be build with gcc-6.

Also adjusted the makefile to pass parameters for the make clean, as some platform specific objects
are not removed otherwise.

gpsp:

Forced gcc-6 as with lr-gpsp.

Simplified the module flags to just include videocore platforms.