system - fix issue with dispmanx flag not being set in chroot environment
This only affected amiberry binaries since Feb 2020 when the build logic was switched
to use this flag. This meant a pure SDL2 version was being built rather than the dispmanx+SDL2 version.

Also corrected the flag setting which was changed in the rpi_aarch64 branch to be an array but this was missed.
The code actually still worked, but it should have been +=(dispmanx) not +=" dispmanx".

__has_dispmanx env var can be used to force this flag for kms/fkms.