ti99sim: use an older version to fix installation with older `gcc`,
Latest version (0.16) needs `gcc` 8 for C++20 features (`--std=c++2a`). This breaks the module for systems with older `gcc` (ex. Debian Stretch).
Added a new scriptmodule based on the previous version (0.15), which used SDL1 instead of SDL2, and keep the module flags from before a50179093.
Added a check for the `gcc` version in the SDL2 scriptmodule.