xroar: add a workaround for the crash on VC platforms
The changes in 25d6710f9275e49 lead to a crash when the emulator starts in fullscreen,
when using the RPI driver in SDL2. Looks like repeated calls to create/destroy the SDL renderer,
when resizing the window, lead to a VC driver error:

     failed to add service - already in use?

The RPI driver already starts the window in fullscreen, so there's no visible difference without the paramter removed.