builder - don't return on error
some modules will return an error due to not having required gcc etc. This would best be handled
with a function to hide modules from building, but in the meantime, it's easier to remove this check
until I have improved the builder scripts somewhat as some binaries were missed when building.

This is also due to changes adding additional support for return codes and error checking, which
happened since the builder code was made.