retroarch - fix saving configuration when config keys exists in an include - #3249
If we have a custom config with

 #include "/opt/retropie/configs/all/retroarch.cfg"

Changing an option that is set in the config above in RGUI and using "Save current configuration" from RGUI will not save that option.
No options that exist in the include will be saved.

This is a revert of upstream change https://github.com/libretro/RetroArch/pull/10524/commits/9eb84728