setup - don't clean up sources on failure unless we are calling from setup gui
This adds a __setup global variable which is used for this (set to 1 during setup_depends).

In the past, sources were not cleaned on error which was an issue for some users when an
upstream source changed causing a custom patch to fail or upstream url code changed. It was
changed so source folders were cleaned up on error which was useful for end users.

This wasn't helpful for development and testing though, so when using retropie_packages
directly, the source folder won't be removed. When building a module and testing via
./retropie_packages.sh MODULE sources, it was  nuisance to have the source folder removed on
a patch failure, especially on larger repositories.