packages - split out / rework install/install_bin logic
We don't want to remove install folder before we download a binary - in the case of a download failure we could end up
just removing the installed version which isn't wanted. Added the removal to before the tar unpacking / verification.

Remove some checks for function existing for cases where it had been done earlier and move some logic for
install/install_bin to later case statement to clean up logic a bit