ppsspp / ppsspp-1.5.4 - added ppsspp 1.5.4 using parent ppsspp module
Users reported that this version runs faster on videocore on the rpi1/2/3

Adjusted ppsspp sources to allow branch parameter and fixed build subfolder to ppsspp
so that paths are the same for md_ret_files, including adjusting lr-ppsspp in the same way.

Older code builds ok with our current ppsspp module script on rpi1/2/3 + videocore gles2,
but not rpi4, which it isn't needed on anyway. This may need changing if we have to update
ppsspp further in the future.