emulationstation-dev - remove functions to avoid multiple calls to edit ES configs
Currently the code for calling frontend hooks just calls all modules with the hook functions
_add_system_* _del_system_* _add_rom_* etc. For additional frontends we check if they are installed in these
functions, but for emulationstation-dev it just called the emulationstation module functions. This actually
meant that we called the emulationstation functions twice when adding a system or removing.

The helper functions probably should do the check to only call the frontend hook functions when they are installed,
which may be a better fix, but for now, this works. I am overhauling some of the packaging code, so I will revisit this.