lr-tyrquake - adjust game_data function to be safer to call as a standlone function
Function could fail if RetroPie-Setup/tmp didn't exist - as we use $__tmpdir primarily for building,
I have switched this to use mktemp -d.

Create destination folders also as required.

I ran into this bug when installing lr-tyrquake from binary install on a new system. $__tmpdir is not actually
created by default, but in this case when unpacking archives, it's better to use mktemp as detailed.