retropie-setup - bump version to 4.7.6
The version was bumped to 4.7.5 just after the 3rd_party_repos PR was merged. However this had a major
bug due to the change in way modules were scanned causing the md_type to contain a full path. It was reverted
minutes after merging, but in case anyone accidentally updated to it, I am bumping again which may help
identify if a user did an update during this short window