packages / helpers - allow rp_module_repo to have function hooks as parameters
rp_module_repo is currently in the format

TYPE URL BRANCH COMMIT

where type is file, git or svn

If rp_module_repo URL BRANCH or COMMIT starts with a : the rest is considered to be a function that
will return the URL, BRANCH, or COMMIT via an echo. This allows for using of the new rp_module_repo
by modules that use some logic to determine what branch / commit is used.

eg. Some modules will use an older branch based on the system they are running on.

These variables are resolved before use, as some modules query the net for information, so we only
want to check these just before they are used.