setup - change colours around to work with other terminals
On the Raspberry Pi console, bold text results in a lighter colour, rather than actually being bold.

One user (it may not be a common issue) reported on their terminal that they couldn't see installed
packages. This was due to our "bold white" for installed applications which ended up being bold grey on a grey
background - so they couldn't see what was installed without hovering over it.

We don't colour installed packages now - but add unavailable packages as red.

This can always be changed later also if needed of course.