packages - add support for module handling of rp_module_repo via hook as first parameter
If the first parameter of rp_module_repo starts with a : it the following function will be called
when setting package information during source install, and when checking for a new source version.

The function takes the parameters "get" and "check".

"get" - should return code to set any packaging variables needed for the module hook later.

The pkg_repo_extra field has been added specifically for modules to store something different from the
standard fields also

"newer" will be called with pkg_repo_* vars set, and can be used by the function hook to determine if
a new version is available. If so, the function should return 0, or 1 if no update is required.

I also split off checking remote repo hash/commit to separate function and fix up an issue with checking
remote git hash, as previously a branch or tag could have multiple matches - we now specifically grep
the results for ref/heads and ref/tags