mupen64plus - rework logic with new rp_module_repo hook to check for updates
Split off the code to get the various repos depending on platform, adding default branch names.

Add a _pkg_info_mupen64plus function which takes get/newer/check parameters from the main packaging code.

If "get" is called, we check all used repositories, getting the last commit hash of each repo, and make
a md5 checksum which is stored in pkg_repo_extra.

If "newer" is called, we check the remote hashes of all repositories, make a md5 checksum and compare against
installed package pkg_repo_extra.

Switched the hash for GLideN64 for older cmake to a full hash, so we generate the same md5sum also.

If "check" is called (only currently used by our tools module for checking repo validity) it checks all the
repos, outputting any errors, and returns 1 if any of them failed.

This is a lot of extra code, but it's only this module which requires this complexity - we could have just
used the main repository, but as some of the plugins are more actively developed at times, it made sense to
want to update if any of the plugin repositories were changed also. This seemed the best way to do it, but
the additional hook functionality could be used with other modules in the future if it was needed.